added tasks.py to ledger app; created user app; moved user model from ledger to user
